project1/requirements.txt
streamlit
google-generativeai
pypdf
pytesseract
pdf2image
Pillow
python-dotenv
sentence-transformers
faiss-cpu
groq